big refactor + pre-commit hooks
