+++ .pre-commit-config.yaml
@@ -8,6 +8,8 @@ repos:
     rev: v4.0.1
     hooks:
     -   id: check-case-conflict
+    -   id: check-json
+    -   id: check-symlinks
     -   id: check-yaml
     -   id: destroyed-symlinks
     -   id: end-of-file-fixer
@@ -37,3 +39,9 @@ repos:
             --check-filenames,
             --check-hidden
         ]
+
+-   repo: https://github.com/pycqa/flake8
+    rev: 4.0.1
+    hooks:
+    -   id: flake8
+        args: ['--ignore=E,F403,F405,F541,F841,W', '--select=E9,F,W6', '--per-file-ignores=__init__.py:F401']
