+++ mii/__init__.py
@@ -1,13 +1,8 @@
-import enum
 import grpc
 from .server_client import MIIServerClient, mii_query_handle
 from .deployment import deploy
 from .terminate import terminate
-from .config import MIIConfig
 from .constants import DeploymentType, Tasks
 
-from .utils import get_model_path, import_score_file, set_model_path, extract_query_dict, mii_cache_path
-from .utils import setup_task, get_task, get_task_name, check_if_task_and_model_is_supported, check_if_task_and_model_is_valid
+from .config import MIIConfig
 from .grpc_related.proto import modelresponse_pb2_grpc
-from .grpc_related.proto import modelresponse_pb2
-from .models.load_models import load_models
