+++ mii/models/__init__.py
@@ -1,0 +1,2 @@
+from .score import create_score_file
+from .load_models import load_models
