+++ mii/models/score/__init__.py
@@ -1,0 +1,1 @@
+from .generate import create_score_file, generated_score_path
